requirements.txt
pandas
numpy
transformers
torch
python-dotenv
google-api-python-client
beautifulsoup4
langdetect
